+++ doc/changelog.rst
@@ -10,6 +10,7 @@ a MongoDB 2.4 server or older, PyMongo now throws a
 
 Highlights include:
 
+- New Session API, see :meth:`~pymongo.mongo_client.MongoClient.start_session`.
 - New methods :meth:`~pymongo.collection.Collection.find_raw_batches` and
   :meth:`~pymongo.collection.Collection.aggregate_raw_batches` for use with
   external libraries that can parse raw batches of BSON data.